Addition of sugar/water ratio
The original sheet did not show the ratio of sugar (g) to water (ml).
Sugar can only be dissolved in water up to 180g in 100ml. So a ratio of 1.8.
The sheet will show you the ratio, which is calculated by the total amount of sugar required and volume of water used.

If the ratio is above 1.8, a warning message will pop-up and tell you to add more water.
I recommend to always aim for at least 2ml of priming solution per bottle (i.e. 50 bottles, 100ml water).
This will ensure that even at 2.5 vols CO2 of table sugar at 20 degrees C, the ratio will be within limits. This is however the case for bottles up to 0.5L. If larger bottles are used, more water might be necessary.