V2-Priming-sugar-Per-Bottle
This version includes the sugar/water ratio, and sends a warning if there is not enough water for the required sugar to dissolve in.

When bottling, ensure that the calculation gives you at least 2ml of solution per bottle. This will ensure an acceptable ratio for bottles up to 0.5L, with all temperatures and CO2 volumes. (when larger bottles are packaged more solution will be required per bottle to keep the ratio in check).